+++ mobile/android/android-components/components/browser/toolbar/src/main/java/mozilla/components/browser/toolbar/display/DisplayToolbar.kt
@@ -25,7 +25,6 @@ import mozilla.components.browser.toolbar.internal.measureActions
 import mozilla.components.browser.toolbar.internal.wrapAction
 import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.concept.toolbar.Toolbar.SiteSecurity
-import mozilla.components.support.ktx.android.content.res.pxToDp
 import mozilla.components.support.ktx.android.view.isVisible
 import mozilla.components.ui.icons.R.drawable.mozac_ic_globe
 import mozilla.components.ui.icons.R.drawable.mozac_ic_lock
@@ -76,7 +75,7 @@ internal class DisplayToolbar(
         set(value) { menuView.menuBuilder = value }
 
     internal val siteSecurityIconView = AppCompatImageView(context).apply {
-        setPadding(resources.pxToDp(ICON_PADDING_DP))
+        setPadding(resources.getDimensionPixelSize(R.dimen.mozac_browser_toolbar_icon_padding))
 
         setImageResource(mozac_ic_globe)
 
@@ -89,7 +88,7 @@ internal class DisplayToolbar(
     internal val titleView = AppCompatTextView(context).apply {
         id = R.id.mozac_browser_toolbar_title_view
         gravity = Gravity.CENTER_VERTICAL
-        textSize = URL_TEXT_SIZE
+        textSize = URL_TEXT_SIZE_SP
         visibility = View.GONE
         ellipsize = TextUtils.TruncateAt.END
 
@@ -99,7 +98,7 @@ internal class DisplayToolbar(
     internal val urlView = AppCompatTextView(context).apply {
         id = R.id.mozac_browser_toolbar_url_view
         gravity = Gravity.CENTER_VERTICAL
-        textSize = URL_TEXT_SIZE
+        textSize = URL_TEXT_SIZE_SP
 
         setSingleLine(true)
         isClickable = true
@@ -331,8 +330,9 @@ internal class DisplayToolbar(
             urlView.measure(urlWidthSpec, fixedHeightSpec)
         }
 
-        val progressHeightSpec = MeasureSpec.makeMeasureSpec(resources.pxToDp(PROGRESS_BAR_HEIGHT_DP),
-                MeasureSpec.EXACTLY)
+        val progressHeightSpec = MeasureSpec.makeMeasureSpec(
+            resources.getDimensionPixelSize(R.dimen.mozac_browser_toolbar_progress_bar_height),
+            MeasureSpec.EXACTLY)
         progressView.measure(widthMeasureSpec, progressHeightSpec)
 
         urlBoxView?.let {
@@ -477,12 +477,9 @@ internal class DisplayToolbar(
     companion object {
         internal const val MEASURED_HEIGHT_THIRD_DENOMINATOR = 3
         internal const val MEASURED_HEIGHT_DENOMINATOR = 2
-        internal const val URL_FADING_EDGE_SIZE_DP = 24
 
-        const val BOTTOM_PROGRESS_BAR = 0
+        internal const val BOTTOM_PROGRESS_BAR = 0
         private const val TOP_PROGRESS_BAR = 1
-        private const val ICON_PADDING_DP = 16
-        private const val URL_TEXT_SIZE = 15f
-        private const val PROGRESS_BAR_HEIGHT_DP = 3
+        private const val URL_TEXT_SIZE_SP = 15f
     }
 }