+++ mobile/android/android-components/components/support/ktx/src/main/java/mozilla/components/support/ktx/android/content/res/Resources.kt
@@ -5,6 +5,9 @@ import android.util.TypedValue
 
 /**
  * Converts a value in density independent pixels (pxToDp) to the actual pixel values for the display.
+ * @deprecated use Int.dpToPx instead.
  */
+// @Deprecated("Use Int.dpToPx instead",
+// ReplaceWith("pixels.dpToPx(this.displayMetrics)", "mozilla.components.support.ktx.android.util.Int.dpToPx"))
 fun Resources.pxToDp(pixels: Int) = TypedValue.applyDimension(
         TypedValue.COMPLEX_UNIT_DIP, pixels.toFloat(), displayMetrics).toInt()