+++ mobile/android/android-components/components/support/ktx/src/test/java/mozilla/components/support/ktx/android/util/DisplayMetricsTest.kt
@@ -1,4 +1,4 @@
-package mozilla.components.support.ktx.android.content.res
+package mozilla.components.support.ktx.android.util
 
 import android.util.DisplayMetrics
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -8,17 +8,16 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class ResourcesTest {
+class DisplayMetricsTest {
 
     @Test
     fun `dp returns same value as manual conversion`() {
-        val resources = testContext.resources
-        val metrics = resources.displayMetrics
+        val metrics = testContext.resources.displayMetrics
 
         for (i in 1..500) {
             val px = (i * (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)).toInt()
-            Assert.assertEquals(px, resources.pxToDp(i))
-            Assert.assertNotEquals(0, resources.pxToDp(i))
+            Assert.assertEquals(px, i.dpToPx(metrics))
+            Assert.assertNotEquals(0, i.dpToPx(metrics))
         }
     }
 }