+++ mobile/android/android-components/samples/toolbar/src/main/java/org/mozilla/samples/toolbar/ToolbarActivity.kt
@@ -33,7 +33,7 @@ import mozilla.components.browser.menu.item.SimpleBrowserMenuItem
 import mozilla.components.browser.toolbar.BrowserToolbar
 import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.feature.toolbar.ToolbarAutocompleteFeature
-import mozilla.components.support.ktx.android.content.res.pxToDp
+import mozilla.components.support.ktx.android.util.dpToPx
 import mozilla.components.support.ktx.android.view.hideKeyboard
 
 /**
@@ -274,12 +274,12 @@ class ToolbarActivity : AppCompatActivity() {
         // //////////////////////////////////////////////////////////////////////////////////////////
 
         toolbar.setBackgroundColor(0xFF38383D.toInt())
-        toolbar.layoutParams.height = toolbar.resources.pxToDp(104)
+        toolbar.layoutParams.height = 104.dpToPx(resources.displayMetrics)
         toolbar.setPadding(
-                resources.pxToDp(58),
-                resources.pxToDp(24),
-                resources.pxToDp(58),
-                resources.pxToDp(24))
+            58.dpToPx(resources.displayMetrics),
+            24.dpToPx(resources.displayMetrics),
+            58.dpToPx(resources.displayMetrics),
+            24.dpToPx(resources.displayMetrics))
 
         // //////////////////////////////////////////////////////////////////////////////////////////
         // Hide the site security icon and set padding around the URL
@@ -288,11 +288,11 @@ class ToolbarActivity : AppCompatActivity() {
         toolbar.displaySiteSecurityIcon = false
 
         toolbar.setUrlTextPadding(
-                left = resources.pxToDp(16),
-                right = resources.pxToDp(16)
+                left = 16.dpToPx(resources.displayMetrics),
+                right = 16.dpToPx(resources.displayMetrics)
         )
 
-        toolbar.browserActionMargin = toolbar.resources.pxToDp(16)
+        toolbar.browserActionMargin = 16.dpToPx(resources.displayMetrics)
 
         // //////////////////////////////////////////////////////////////////////////////////////////
         // Add a custom "URL box" (url + page actions) background view that also acts as a custom
@@ -302,7 +302,7 @@ class ToolbarActivity : AppCompatActivity() {
         val urlBoxProgress = UrlBoxProgressView(this)
 
         toolbar.urlBoxView = urlBoxProgress
-        toolbar.urlBoxMargin = toolbar.resources.pxToDp(16)
+        toolbar.urlBoxMargin = 16.dpToPx(resources.displayMetrics)
 
         // //////////////////////////////////////////////////////////////////////////////////////////
         // Add navigation actions
@@ -390,7 +390,7 @@ class ToolbarActivity : AppCompatActivity() {
         // Add a fixed size element for spacing and a branding image
         // //////////////////////////////////////////////////////////////////////////////////////////
 
-        val space = Toolbar.ActionSpace(toolbar.resources.pxToDp(128))
+        val space = Toolbar.ActionSpace(128.dpToPx(resources.displayMetrics))
 
         toolbar.addBrowserAction(space)